@@ -3,7 +3,6 @@ import 'dart:convert' as convert;
 import 'package:account_app/entity/account.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:http/http.dart' as http;
 
@@ -18,9 +17,9 @@ class InputFormState extends StateNotifier<RadioValue> {
 }
 
 final inputFormProvider = StateNotifierProvider<InputFormState, RadioValue>(
-  (refs) => InputFormState());
+    (refs) => InputFormState());
 
-class InputForm extends HookWidget {
+class InputForm extends ConsumerWidget {
   static Route<dynamic> route() {
     return MaterialPageRoute<dynamic>(
       builder: (_) => InputForm(),
@@ -31,13 +30,13 @@ class InputForm extends HookWidget {
 
   final Account _data = Account(0, 0, '', 0);
 
-  BuildContext _context;
+  BuildContext? _context;
 
-  RadioValue _groupValue;
+  RadioValue? _groupValue;
 
   @override
-  Widget build(BuildContext context) {
-    _groupValue = useProvider(inputFormProvider);
+  Widget build(BuildContext context, WidgetRef ref) {
+    _groupValue = ref.watch(inputFormProvider);
     _context = context;
     return Scaffold(
       appBar: AppBar(
@@ -46,40 +45,43 @@ class InputForm extends HookWidget {
       body: SafeArea(
         child: Form(
           key: _formKey,
-          child: ListView(
-            padding: const EdgeInsets.all(20),
-            children: <Widget>[
-              _createRadioListColumn(context, _groupValue),
-              _createTextFieldColumn(),
-              _createSaveIconButton(_formKey),
-            ],
+          child: SizedBox(
+            child: ListView(
+              shrinkWrap: true,
+              padding: const EdgeInsets.all(20),
+              children: <Widget>[
+                _createRadioListColumn(ref, _groupValue!),
+                _createTextFieldColumn(),
+                _createSaveIconButton(ref, _formKey),
+              ],
+            ),
           ),
         ),
       ),
     );
   }
 
-  Widget _createRadioListColumn(BuildContext context, RadioValue groupValue) {
+  Widget _createRadioListColumn(WidgetRef ref, RadioValue groupValue) {
     return Padding(
       padding: const EdgeInsets.all(0),
       child: Column(
         children: [
-          _createRadioList(context, groupValue, RadioValue.INCOME),
-          _createRadioList(context, groupValue, RadioValue.SPENDING),
+          _createRadioList(ref, groupValue, RadioValue.INCOME),
+          _createRadioList(ref, groupValue, RadioValue.SPENDING),
         ],
       ),
     );
   }
 
   Widget _createRadioList(
-      BuildContext context, RadioValue groupValue, RadioValue value) {
+      WidgetRef ref, RadioValue groupValue, RadioValue value) {
     final text = value == RadioValue.SPENDING ? '支出' : '収入';
 
     return RadioListTile(
       title: Text(text),
       value: value,
       groupValue: groupValue,
-      onChanged: (value) => _onRadioSelected(value, context),
+      onChanged: (value) => _onRadioSelected(value, ref),
     );
   }
 
@@ -102,11 +104,11 @@ class InputForm extends HookWidget {
         hintText: '項目',
         labelText: 'Item',
       ),
-      onSaved: (String value) {
-        _data.item = value;
+      onSaved: (String? value) {
+        _data.item = value!;
       },
       validator: (value) {
-        if (value.isEmpty) {
+        if (value!.isEmpty) {
           return '項目は必須入力項目です';
         }
         return null;
@@ -122,11 +124,11 @@ class InputForm extends HookWidget {
         hintText: '金額',
         labelText: 'Money',
       ),
-      onSaved: (String value) {
-        _data.money = int.parse(value);
+      onSaved: (String? value) {
+        _data.money = int.parse(value!);
       },
       validator: (value) {
-        if (value.isEmpty) {
+        if (value!.isEmpty) {
           return '金額は必須入力項目です';
         }
         return null;
@@ -135,18 +137,18 @@ class InputForm extends HookWidget {
     );
   }
 
-  Widget _createSaveIconButton(GlobalKey<FormState> formKey) {
+  Widget _createSaveIconButton(WidgetRef ref, GlobalKey<FormState> formKey) {
     return Padding(
       padding: const EdgeInsets.all(30),
       child: ElevatedButton(
         onPressed: () {
-          formKey.currentState.save();
+          formKey.currentState!.save();
           _data.type =
-          _context.read(inputFormProvider.notifier).state == RadioValue.INCOME
-              ? 1
-              : 0;
+              ref.read(inputFormProvider.notifier).state == RadioValue.INCOME
+                  ? 1
+                  : 0;
           saveHouseholdAccountData(_data);
-          Navigator.of(_context).pop<dynamic>();
+          Navigator.of(_context!).pop<dynamic>();
         },
         child: const Text('登録'),
       ),
@@ -155,7 +157,7 @@ class InputForm extends HookWidget {
 
   Future<void> saveHouseholdAccountData(Account data) async {
     final uri = Uri.parse(
-      'https://script.google.com/macros/s/AKfycbwqSD3jNHnUvG30N0CKJyLEwqTtdRCs9ewuVTHDAOqf3dzea7_L/exec');
+        'https://script.google.com/macros/s/AKfycbwqSD3jNHnUvG30N0CKJyLEwqTtdRCs9ewuVTHDAOqf3dzea7_L/exec');
     final headers = {
       'Content-Type': 'application/x-www-form-urlencoded',
     };
@@ -182,7 +184,7 @@ class InputForm extends HookWidget {
     }
   }
 
-  void _onRadioSelected(dynamic value, BuildContext context) {
-    context.read(inputFormProvider.notifier).changeState(value);
+  void _onRadioSelected(dynamic value, WidgetRef ref) {
+    ref.read(inputFormProvider.notifier).changeState(value);
   }
 }
